+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Add `fail_on_duplicate_payment_method_for_customer` option to 
+  - `ClientToken`
+  - `PaymentMethod`
+  - `CreditCard`
 - Add `blik_aliases` to LocalPaymentCompleted webhook
 - Deprecate `samsung_pay_card`
 - Updated expiring pinned vendor SSL certificates